move VALID_REGIONS

+++ lib/aws_s3_client.rb
@@ -1,6 +1,24 @@
 # frozen_string_literal: true
 
 class AwsS3Client
+  VALID_REGIONS = %w[
+    ap-south-1
+    ap-northeast-2
+    ap-southeast-1
+    ap-southeast-2
+    ap-northeast-1
+    ca-central-1
+    eu-central-1
+    eu-west-1
+    eu-west-2
+    eu-west-3
+    eu-north-1
+    sa-east-1
+    us-east-1
+    us-east-2
+    us-west-1
+    us-west-2
+  ]
   attr_reader :client, :resource
 
   def initialize(settings)
@@ -15,6 +33,8 @@ class AwsS3Client
   end
 
   def new_bucket(params)
+    raise 'Request invalid!' unless valid_bucket_params?(params)
+
     client = Aws::S3::Client.new(region: params['bucket_region'], credentials: @credentials)
     result = client.create_bucket(
       bucket:                      params['bucket_name'],
@@ -36,6 +56,8 @@ class AwsS3Client
   end
 
   def delete_bucket(params)
+    raise 'Request invalid!' unless valid_bucket_params?(params)
+
     client = Aws::S3::Client.new(region: params['bucket_region'], credentials: @credentials)
     client.delete_bucket(bucket: params['bucket_name'])
   end
@@ -69,5 +91,9 @@ class AwsS3Client
 
   def credentials(settings)
     Aws::Credentials.new(settings[:access_key_id], settings[:secret_access_key])
+  end
+
+  def valid_bucket_params?(params)
+    !params['bucket_name'].to_s.empty? && VALID_REGIONS.include?(params['bucket_region'])
   end
 end
